Use native screen resolution, and set RESIZABLE flag
Setting 0,0 will create the screen at the full-size of the current
screen. By setting RESIZABLE, it makes it possible to resize the window
once it’s been created.

This doesn’t work perfect yet. The buttons don’t yet scale with the
screen, and if you resize the screen too small it could crash.